Create banking.py
one of projects in JetBrains -
i tryed to make difference between terminal, and bank system as main part (here are some not finished parts for ex. printing some messages in functions inside bank procedures. how i understand it must be returned by functions and then printed inside terminal procedures) but... it is work...

and parced tests... so, let's go for the next task...